# Mayerj/optimize duplicate code