--- a/Philips.CodeAnalysis.sln
+++ b/Philips.CodeAnalysis.sln
@@ -20,6 +20,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Philips.CodeAnalysis.Mainta
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Philips.CodeAnalysis.DuplicateCodeAnalyzer", "Philips.CodeAnalysis.DuplicateCodeAnalyzer\Philips.CodeAnalysis.DuplicateCodeAnalyzer.csproj", "{378CCB57-00DA-4782-A238-ADC134827B93}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Philips.CodeAnalysis.Benchmark", "Philips.CodeAnalysis.Benchmark\Philips.CodeAnalysis.Benchmark.csproj", "{D9B847FB-E395-4875-86E9-C69B47B8D77C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -50,6 +52,10 @@ Global
 		{378CCB57-00DA-4782-A238-ADC134827B93}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{378CCB57-00DA-4782-A238-ADC134827B93}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{378CCB57-00DA-4782-A238-ADC134827B93}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D9B847FB-E395-4875-86E9-C69B47B8D77C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D9B847FB-E395-4875-86E9-C69B47B8D77C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D9B847FB-E395-4875-86E9-C69B47B8D77C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D9B847FB-E395-4875-86E9-C69B47B8D77C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE