This has several (small) performance improvements to the duplicate code analyzer.  Things like pre-allocating a list if we know the intended length, not creating the LocationEnvelopes until they are necessary, etc.

It also includes a change to the rolling hash algorithm, increasing the bucket count enormously.  On large projects the small bucket count was causing large numbers of collisions, which are extremely slow to resolve.

From the benchmarking: (2 input projects, one a "small" amount of code, one "large").

|                           Method |                    A |     Mean |    Error |   StdDev |
|--------------------------------- |--------------------- |---------:|---------:|---------:|
|                         OriginalHashParameters | C:\Ph(...)vices [56] |  4.545 s | 0.2088 s | 0.1630 s |
|             BiggerPrimes | C:\Ph(...)vices [56] |  1.122 s | 0.1357 s | 0.1203 s |
|                         OriginalHashParameters | C:\Ph(...).Auto [62] | 26.736 s | 0.8556 s | 0.7144 s |
|             BiggerPrimes | C:\Ph(...).Auto [62] |  3.344 s | 0.7255 s | 0.6432 s |

See [mayerj/performance-test](https://github.com/philips-software/roslyn-analyzers/tree/mayerj/performance-test) for more performance tests